Refactor FindDeclarationVariableNodes and move helper methods out to DeclarationVariables class
